pipeline/requirements_pipeline.txt
pandas>=1.5
numpy>=1.24
scikit-learn>=1.2
xgboost>=1.7
imbalanced-learn>=0.10
matplotlib>=3.6
seaborn>=0.12
joblib
tensorflow>=2.12  
tqdm
polar
